Show series ACL in uploader when configured
And then also disable ACL editing and attachment.
In a previous iteration, this was using a suboptimal
approach where each series returned by the API would
send an additional query to get the series ACL.
Now this only sends a single additional query whenever
a series is actually selected. While this requires
more code, a good amount of it is just state management
and error handling/reporting. The rest is related to
the data fetching.