Use Meili's federated search instead of manually merging results
This has several advantages:
- More well defined ordering, with simple way to boost some item types.
- Easier to implement pagination.
- Slightly less code in Tobira
- This could be faster (I haven't really observed a notable difference
  in my tiny tests though.

This might change the result set for certain queries. I just quickly
checked several cases to make sure it's still useful. Returning the
exact same results as before is not important, it should just be useful.
And the idea is that with this, it can be more useful than before.