Boost series weight in main search
This is is mostly done for one common case: a series where all events
have the same title as the series. Searching for the series name
previously showed all videos (in pretty much random order) first, and
only then the series. With this tiny boost, the series is shown first
in these cases. You can try that with the query "Quantenelektronik".