Generalize event table backend logic
This generalizes and factors out some backend code
so it can also be used for series and playlists.

It also replaces the cursor based pagination with an
offset based one. That simplifies a lot of things and allows
us to get rid of a bunch of code.

The frontend now uses a `page` url parameter to signify the current
page. When entered manually, there are some checks and logic to make
sure it's always in bounds, i.e. using a number smaller than 0 will
always redirect to `page=1` and a number larger that the max number
of pages will redirect to the last page.

Please note that the generalization of the sorting columns will be done
in a later commit.