Move `isTransitioning` into `RouterState` to avoid useless renders
The `isTransitioning` state was part of the router context meaning that
all components subscribed to the context were rerendered when the state
changed. That is undesirable as this caused many useless renders. One
could still keep one context and libraries like `use-context-selector`,
but that seemed overkill for this one usecase. I also considered passing
`LoadingIndicator` to `Router` which works, but `header/Search` also
uses `isTransitioning`. This will likely be removed when I removed
instant search, but to keep this a well contained PR and for future
flexibility, I opted for the "two contexts" solution.