Memoize rendering of route content in `RootLoader`
There are still some renders that seem useless to me but that I cannot
find a way to get rid of. It's infuriating really. This change at least
makes sure that the expensive part (the actual page content) is not
rendered uselessly. Again, it would be nicer to make sure, `ActiveRoute`
and `RootLoader` aren't rendered uselessly in the first place, but I
couldn't make that happen.