Fix `SearchableSelect` items reusing thumbnails
I always noticed some weird behavior but finally checked it with a
screen capture. And indeed, when you input a new character and if the
search returns quickly enough (to avoid showing "loading"), then the
new list reuses old <li> items. Specifically: the text data is updated,
but the `<img>` is reused and it looks like only the `src` is updated
by react. The effect of this is that the browser first keeps showing
the thumbnail of the event that was previously at that spot, until the
new image is loaded. By properly setting `key`, this is fixed.