Add API for changing video/series metadata
This adds a somewhat generic function that can be used
to send `update metadata` put requests to Opencast for
both events and series, as their respective endpoints
only differ in one string (i.e. `series` vs `events`).

For now, the function is only used to update title and
description, but it can be used for any kind of dublincore
metadata.