Add search-match simplification algorithm to limit timeline complexity
Previously (especially before using stop words), the search endpoint
could return a vast number of text matches, slowing down network and
frontend. This slow down needs to be avoided and, in those cases, seeing
all the individual matches is not useful anyway. So we reduce the number
of returned matches to a limit, without changing the "general look of
the timeline" too much. See comments for more details.