Fix known groups/users label-related default key and typing
Things I forgot in #1316. The reason for why I didn't notice is that the
`TranslatedLabel` type was bad. In particular, `Record<string, _>` seems
to behave differently than I assumed. It's very lenient. Now I made
typing more strict in multiple places. In the backend, the translated
string will also check the default field when loading from DB, in order
to keep the API description guarantees.

This fixes the ACL editor being buggy or crashing when dealing with
users.