Adjust query and `LoadableAsset` to include deleted events
I forgot the fact that the table might include events that
have been marked as deleted. Unfortunately that means the
query and `LoadableAsset` trait are getting increasingly
complicated to the point where it's hard to understand and
quite easy to make these errors.
This is also in part due to the series table's need to be
sortable by event count, and I am starting to wonder if it
might be better to move the whole query to the trait, though
that brings its own complications and I doubt that would even
be possible without other major adjustments.