Add migration, API and mutation for series deletion
This repurposes the `delete` endpoint previously used
by events, and basically mirrors everything done when
an event is deleted.

This means the series table is renamed to `all_series`
and gets a field `tobiraDeletionTimestamp`, which is
used to mark a series as deleted. The original series
table is moved to a view.

When a series is deleted in Tobira, the timestamp is saved
and series marked as deleted, but still present in Opencast
get a special entry in the `My series` table. Every other
place uses the series view without deleted series.

When the series is actually deleted in Opencast and removed
from the DB with sync, the entry from the `My series` table
also disappears.