Add series deletion related UI features in frontend
This, once again, is a bunch of generalization to leverage
the fact that we already have basically all that is needed
for this.
While that makes it sound relatively simple, this was still
a lot of work, mostly to not just copy everything but factor
out as much as possible and make it usable for both events and
series, and possibly also playlists in the future.
There are still some things that need ironing out:
- There is a deserialization error where postgres complains
about a null value
- The translation/i18n doesn't seem to be working correctly for all
cases