Avoid duplicate `blocks` SQL query for `RealmQuery`
`blocks { ... }` were included in the main query and the `BlocksData`
fragment. For some reason, this causes Juniper to execute the SQL query
loading block data twice. That's not great. An easy fix is to just
remove it from the main query. This requires us to move the
`WelcomeMessage` to `blocks/index.tsx` but that's not a big deal.