Optimize `resolved_name` querying in realm SQL queries
The `resolved_name` is an SQL function. Unfortunately, using it makes
the query a lot slower. The previous query:

    select id, realms.resolved_name from realms where id = 0

... takes between 250µs and 350µs. This change improves the runtime to
roughly 15µs for realms that do not have a derived name. Of course it
is duplicated code, but it's not that bad I think. It won't change
often.

I also removed the manual selection in `Realm::root`. There is really no
point anymore in trying to reduce DB traffic a bit by not selecting some
fields we already know for the root realm. Back in the day we didn't
need to do any DB query as everything about the root realm was already
known. But over time we needed to load more and more data from the DB
so no point in not using the `FromDb` impl.

Here is some more detailed data for the whole runtime of page queries:
- Root realm: 750µs -> 240µs
- Non-root realm: 950µs -> 440µs
- Video page in root: 700µs -> 200µs
- Video page in non root: 900µs -> 400µs

Very rough numbers but yes, the `resolved_name` is actually a big part
of the SQL runtime for common GQL queries.