Fix series thumbnails on `SeriesDetails` page
This page uses the `series_by_id` api, which doesn't
use the new `thumbnailStack` field. It could, but that would
need some additional mapping in backend and should then
probably also be used in all other places that need
a series thumbnail.
That would have the added benefit of these thumbnails always
having the same order everywhere. Right now, that's not given
since the only place the `thumbnailStack` is used as of now
orders them in backend by their creation date (ascending),
which we don't do for the other occurences that are mapped in
frontend.