Make JWT contain auth claims to grant correct permissions
Previously, our JWTs just contained user information, but not actual
information about what the user is supposed to be able to do in
Opencast. The roles for JWT users were just hard coded in the Opencast
security config. But this did not quite work for the editor, as there,
Tobira has to state "this user can edit this event" (see #600).

This commit fixes that by attaching auth information according to the
standard introduced here: https://github.com/opencast/opencast/pull/6177

For the uploader and studio, `role: ["ROLE_STUDIO"]` is passed. For the
editor, that's `oc: { "e:${id}": ["read", "write"] }`.

This commit also passes the username as `sub` in addition to `username`,
to be in line with said standard. Including the username twice is a
waste of space of course, but for now it's the best way to be backwards
compatible with older Opencasts. We will remove the `username` claim
in the future, which means admins need to update their JWT config in
Opencast.