Add/modify frontend components to allow series video management
The core component of this is the table for adding and removing
videos from video lists. This is written in a way that should
allow it to also be used for playlists later on, by passing
a custom base mutation - though this will probably still need
further adjustments to make that truly the case. But these
adjustments should be made when that playlist feature is actually
getting implemented.

For series video management I had to add two new
query parameters to the video list selector. These can be used
to (a) set a boolean which when set will filter out all events
that are already part of a series, and (b) pass a list of
event ids that will also be filtered out (these are events that
are being added to the list but not submitted yet).