Make event title and description editable
Most of the work for this was already done for
series, so this was a pretty simple addition.
There is some duplication involved, especially
in backend, but I think it wouldn't really be that
practical to factor out these endpoints. There
are small but meaningful differences like response
code (why, Opencast?), return type and the need
to start `republish metadata` for events.