Deduplicate some error messages
These messages were nearly the same and a while a
generalization loses some specificity, that isn't
a big deal imo. It's not that helpful to be told
"changing ACL didn't work" instead of "This action
didn't work" when I just tried to change the ACL and
the error message is shown on that same page.