Stop using ubuntu-20.04 in GitHub actions
The ubuntu-20.04 runner are deprecated for a while now and will be
completely disabled very soon.

For some actions, the runner OS really didn't matter. They changed to
`ubuntu-latest`. But when actually building the Tobira binary, the OS
determines the minimum-required libc version. Previously, our builds
required libc 2.31, but with Ubuntu 22.04, it requires 2.35. Our test
deployment VM uses CentOS Stream 9, which has an older libc version.

With this commit, builds for the test deployment are always -musl
builds. This means it easily works on the test deployment. Further,
release builds will use Ubuntu 22.04 and thus require 2.35 now.

With more work and slowing down CI, we could lower the required libc
version. But given that we offer a -musl build and I cannot see any
disadvantages of that, we don't want to spend any time improving the
-gnu build now.