Make series thumbnail order consistent
There was a mismatch between the order in thumbnail
stacks on the series table, where they were ordered
by their creation date, and their order in search results
and on the series details/acl pages.

Adapting the table order on the management pages is
simple but it's another thing with search results.
For this we would need to write a new db migration,
which is certainly possible but a little annoying.

So for now, this just removes the ordering on the
series table, even though I would rather prefer
to apply it in the other places. TBD.