Add creator UI to series & playlist blocks, date to series blocks
Playlists do not have a `created` date, so we don't show a date for now.
We also decided to simply show this metadata whenever the block has the
`show_metadata` toggle set. Problem is that previously, that checkbox
was labelled "Show description". And maybe users want to separately
control these things? But for now we decided to go the easy path and
tie them to one another.

Also, this currently grabs the creator from `metadata`, which is not
ideal. We should have a API for `creator`, store it separately in the
DB and the harvest API should list it explicitly. But: we have this
`metadata` field for a reason: as a catch all thing. So we might as
well use it here instead of doing all the work required to do it
differently. That will happen anyway when there is a data model change
in Opencast.