Sort series in series selector by `created` instead of `updated`
This unfortunately requires an index rebuild, because we were very
stingy with our index definition before and didn't declare
`created_timestamp` as sortable attribute.

Also, the results are always sorted by that date now, not only when the
query is empty. The previous condition was there because I still wanted
the results to be sorted by relevancy. But Meili values relevancy higher
than the "sort" attribute anyway, so this is pretty much exactly what
one expects. All the results that have roughly the same relevancy are
sorted, less relevant results still come later.