Replace series `syncedData` props with `state`
We previously used the series description as `syncedData`
to determine a series' sync status. Since we now allow
creating series in Tobira (using placeholders util they are
synced with Opencast), this practice isn't viable anymore.

This now uses the series `state` property, which is either
`waiting` or `ready` (with `ready` being synonymous with `synced`).