Improve realm nav by showing siblings when there are no children
Before, the navigation of leaf realms was fairly useless, only allowing
you to go up, never showing more links, making browsing less inviting.
With this commit, the siblings are shown instead in these situations.
To visually distinguish this "special last level" from the others,
in intermediate layers (where children are shown), all children are
indented a bit. Also, the "current realm" name is not shown as second
line in the nav on the last level (no children), as that would show
the realm name twice.

The API and backend was also changed as otherwise, the frontend would
have needed to request the children and siblings all the time, adding
more traffic and another SQL query. The API `Realm` got a new field
`nav` that delivers exactly the information needed to show the
navigation, loading everything in one SQL query (which is one less than
before).