Redesign navigation & show realm siblings when there are no children
Before, the navigation of leaf realms was fairly useless, only allowing
you to go up, never showing more links, making browsing less inviting.
With this commit, the siblings are shown instead in these situations.

To make this work neatly, the navigation was generally redesigned a bit.
Items can be indented and can show tree-icons. The active element looks
a bit different now. And there was lots of refactoring, cleaning up
useless code and working towards #840.

The API and backend was also changed as otherwise, the frontend would
have needed to request the children and siblings all the time, adding
more traffic and another SQL query. The API `Realm` got a new field
`nav` that delivers exactly the information needed to show the
navigation, loading everything in one SQL query (which is one less than
before).