Allow changing built-in group info (e.g. label of ROLE_USER)
Luckily, this didn't require that much work. The database did not
disallow these entries in `known_groups`, neither does the subcommand
`known-groups upsert`. So it was simply a matter of making the frontend
use the custom info if it is provided. The `acl` API docs were also
adjusted to match the actual behavior. It is not quite possible to
override info for ROLE_ADMIN yet, as that special group has quite a few
extra behaviors attached to it, but ROLE_USER and ROLE_ANONYMOUS can be
defined as any other group.