Fix caching header for JS bundle files
When I updated `reinda`, we started using the built in hashes by webpack
but I forgot to update the piece of code setting the caching header.
This means that browsers could not cache that big JS bundle, which is
uhg... kind of annoyed at myself that this happened and stayed unnoticed
for so long. But now this change is an easy performance win.