Add upload button to series video management
The button links to the upload page and locks in the series for the upload,
using an url parameter. This means we need to do some extra datafetching
on the uploader side, but I suppose that is reasonable for that purpose
and won't add too much additional load.
When the parameter is used, the series cannot be changed, and when
`lockAclToSeries` is enabled, the series ACL is shown in place of the
event's ACL, as usual.