Add creation date to series placeholders
Not showing a creation date for items that
have clearly been created is a little confusing.
So even though it will be overwritten with the
date from Opencast when sync happens, showing
a "created in Tobira" date before that should be
fine. It might even be helpful if creation on the
Opencast side failed. Then the user can at least
see when they attempted to create the item.