Improve workflow related handling of Opencast API errors
Previously, when a user tried to visite a page that checked
for workflow activity, the whole page would "crash" and result
in an unrecoverable error. This behaviour has been replaced
by simply showing a note that the OC status could not be obtained
and therefore, editing has been disabled.
This is implemented for the video "details" and "access" pages.