Make harvest code able to deal with full batches of unreadable items
Before, this case lead to an error, being unable to continue harvesting.
But this can actually happen when a full batch of OC items are
unreadable. These are items that are somehow broken and are skipped by
Opencast. They should be rare, but it can happen. This is mostly
relevant for local development, where wiping a local Opencast improperly
(e.g. not clearing the index correctly) causes lots of broken events.