Add `role="contentinfo"` to footer
Note that this should NOT be necessary.
In a recent accessibility review, it was reported
that the footer region, while correctly displayed
in the landmark navigation tree and also labelled as
"contentinfo", is not being read out loud by
the JAWS screenreader (they might need a bigger boat).

On MacOS, I can only test with voiceover, and with that
I cannot confirm the reported issue, so it's kinda hard
to find a fix for this. Adding the redundant `role="contentinfo"`
prop might not help either since, as noted, it should be
and from my end IS already present and implied by using
the html footer element. But it's the only thing I could
think of. In firefox, this doesn't change anything, while
chrome will change the read and displayed text for voiceover
from "footer" to "content info", so I guess it makes at least
some difference.