Make video buttons a section, add some aria labels
Once again, Firefox doesn't do anything with the fact
that it's a section, but chrome will and tell voiceover
that it's a landmark.

Unrelated, but aria labels and a live region for the
timeinput part of the share menu were also added in this
commit. This should help visually impaired users that
rely on screenreaders get some information about the interface
and currently set time.