Perform write-permission check for preselected series in uploader
At the end of the day, this is just a frontend check and can thus be
overcome. But Tobira cannot make this completely save, as Opencast
itself accepts uploads into a series that the user has no write access
to. This check is just added here because the `series` URL parameter
has made it quite a bit easier to abuse this behavior. So this commit
makes it harder again.