Reshuffle react hook form usage in edit mode forms
There was an issue where the `isDirty` form state
wasn't being picked up correctly. This took a look
of research, investigating and ultimately trial and
error to arrive at the pretty simple solution (in relation
to the invested time anyway).

The form is now insantiated on a lower level and get's
passed the respective block default values, with which
the `isDirty` state is now being tracked correctly.