Set TypeScript `target` to `esnext`
`target` determines `lib`, which in turn determines
which library functions TypeScript lets you use without warning.
This is nice and all, but TypeScript actually doesn't know
what our target browsers support.

Babel knows it, and could automatically inject polyfills,
but those would pull in `core-js`, which is not ideal.

Unfortunately it's also not that easy to bridge the gap
between `browserslist` and `tsconfig.json`'s `target`.
This commit basically gives up on it for now,
but see the next one for a potential solution.

This also remove a (now) false-positive `@ts-expect-error`.
See also previously #1471.