Add `eslint-plugin-compat`
This adds an ESLint plugin that can check whether we use
any Web- or ES-API not supported by browsers in our `browserslist` config.

This is not perfect, since it is browser-specific,
but we also lint/typecheck (some of) our config files.
We will probably need to make some more distinctions between those down the line,
but for now, this should already improve the situation.