Fix embed route by moving video <section> out of `VideoContextProvider`
This was added in d20e9196d722acd11382446a765e465dd1e75466 for
accessibility reasons. Unfortunately, this broke embed routes as
the `<section>` would have a height of 0 for some reason, making the
player invisible. One option would have been to add `height: 100%`.
But a component called `...ContextProvider` is usually expected to not
render any real DOM, but just provide a context. So I moved the section
out of it into the two places where it's used. Embed routes don't have
this section now, but I think the accessibility argument doesn't matter
there as it's the only thing on page.