Add service worker injecting JWTs in OC requests (MVP)
This is the first working version of automatically adding JWTs to OC
requests that need authentication. It uses a service worker which can
intercept all requests done by the page. This is useful for a few
reasons:

- All the JWT logic is in one place and doesn't have to be scattered
  around the whole codebase, wherever an OC URL is used.

- This works "just in time", so we can easily make sure to have a
  fresh token. Otherwise, it would get quite tricky! With our `<img>`
  that have `loading="lazy"`, we would have to manually use
  IntersectionObserver to fetch a JWT in time and change the `src`
  attribute. For videos, Paella would need support to hook into several
  events and then change the URL accordingly (e.g. when changing
  resolutions, subtitles tracks, ...). Even worse: since the video is
  loaded progressively over a potentially long time, we need to generate
  new JWTs for later requests. Service worker just cleanly intercept
  the requests when they occur, even working for Range requests.

- Service worker can use the `Authorization` header instead of attaching
  `?jwt=` to the URL. That in turn means that the normal browser cache
  is still used (because the JWT is not part of the cache key).

However, service workers also have disadvantages. They are finicky to
deal with and debug. Their life-cycle is also different from what most
web devs would expect, thus being confusing. Finally, having JS run for
each request can make everything slower. In particular, the use of the
`Authorization` header means that each requests requires a preflight
request.

I will deal with these shortcomings soon, by fully making sure that the
life-cycle is correctly managed here and that updating the SW works as
expected. Further, there might be ways to speed up routing for SW. And
the impact of preflight requests I will measure as well.

The `crossorigin` value of the Paella `<video>` element was set to
"anonymous" which has the same meaning as the `false` before, only that
there is a Firefox bug with the previous value where service workers
are not triggered when loading video data.
See https://stackoverflow.com/a/79062769/2408867

This is not yet shippable! It needs to be configurable and the number of
API requests and generated JWTs can still be reduced a lot by caching
JWTs for example.