Add `auth.hide_login_button` config
This is useful for use-cases where normal users are not supposed to
login and it's expected that only a handful of admins log in. Those
admins can type the direct link to the login page. And showing a login
button kind of suggests to normal users that they can login, which is
confusing if they in fact cannot.