Add scope to player "space" shortcut
This makes sure that the default behaviour
(i.e. scroll by pressing the spacebar) will only
be prevented if there is an active video player on
the current page (scrolling is still prevented in
other standard cases, for example when a button is focused
– this will also not trigger the shortcut).

It should be noted that once a player is active, scrolling
with space will no longer be possible at all on the current
page instance.
Also, and this is definitely a point for further discussion:
Shortcuts will be activated once the player is loaded.
That means the user has to click or otherwise trigger
the play button first before they can use any shortcuts. This
might be annoying to some (I personally do not like that necessity).