Add loadtest scenario for new multi JWT API
On my machine with ED25519, that's ~30k req/s. With each req taking very
roughly half a ms. That's plenty. Especially considering that this
creates 6 JWTs each. From what I can tell (by some quick testing), the
main time sink is actually DB delay. Not even the query, that should run
in 30µs or so (at least without load). And signing should also only take
in the neighborhood of 50µs total for 6 JWTs.

I also built a custom JSON API to avoid GraphQL. But that didn't seem
to make a difference. Which is a bit surprising but hey. I'm fairly
sure that GraphQL does not add a significant slowdown here, which is
great as we can just continue using GraphQL, even for this.