Include all supported languages in group label translations
Adding new languages caused some weird issues with the built in
groups detection. I won't pretend that I completely (or even remotely)
understand what's happening under the hood (not for a lack of trying),
but it has something to do with i18n and local storage.
I think. Maybe not. Maybe it's something else entirely.

So while I'd like to implement a fix that addresses the core issue,
the one that I supply here will have to do for now.
That fix is to not use `i18n.languages` but just use our hardcoded
`languages` object and use its keys to always have an array of all
languages that are usable in Tobira. \ramblemode off\