Re-add condition for showing unlisted note
This got mistakenly dropped during a refactor, so the
note was shown for all videos, regardless of being listed
or not. This restores the former logic of checking if the
video has any "host realms" (which include all blocks
containing the video, i.e. not just video blocks but also
series and playlist blocks).