Add way to overwrite user realm path via auth integration
Before, we always used `/@${username}`, but this makes it possible to
overwrite that handle with the auth integration. At least when using
auth callbacks. While it would have been easy to implement it for auth
headers as well, those are kind of soft-deprecated and we don't want
people to use them.

The new field in the auth callback is `userRealmHandle`. When the auth
integration changes the userRealmHandle for a user that already created
a user realm, the behavior could be confusing. In that case, the realm
needs to be changed in the DB manually.

A previous remote error translation (username-not-valid-as-path) was
broken: it had a {{username}} in it, but that was never set. That
placeholder was removed from it and the text generalized. The IT and FR
translations were removed since I cannot adjust them. Note however that
this error message might not be shown at all anyway. In order for a
user to create a user realm, they must have been on the correct path
already. But if the handle is invalid as a path, they cannot. So as far
as I know, this error can only be triggered by manually sending API
requests. So I don't see a pressing need to translate that error again.