Add config to disable series video removal
This is a temporary solution for when Opencast is
configured to require that videos are always part
of a series, since that would prevent video removal
anyway. We will adjust the UI to make it possible
to directly move a video to another series, but until
then, it's easier to just disable that feature.