Add shortcut to hide/show player ui
The UI can now be hidden and subsequently
shown again by pressing `h`. When hidden,
any mouse activity inside the video container
will also pop the UI back in, preserving the
default behaviour.

You might wonder why this doesn't use built in
paella functions like `hideUserInterface()`.
That only works when the player is not paused,
and even then not reliably. So this implements
a css solution instead.

Note that *showing* will fail initially when the UI
was hidden automatically, because this checks the
state of a css class to track the hidden state.
But even in that case, which I don't expect to be
a common use case, pressing the button twice will
produce the desired effect.