Override paella plugin translations
The german plugin translations aren't working
*on deployments*. They do locally.
I don't exactly know what's causing this, but
we can just specify these manually and adjust
some wording as well. This way, they should
also work when deployed.
Note that we do not use all of these, but seeing
as they're specified in italian and french (which
is also my "mistake"), I just added everything for
the full monty. I believe that is a thing people
say.....