Add admin dashboard
Fixes #475

This adds a page that only proper admins can see, showing an overview
of the most important stats and states of Tobira. It should help
identify problems and show at a glance if Tobira is running well. As
this is only semi-user facing, the page does not look as good as other
pages. Function over form. It is also expected that we add and change
that dashboard regularly.