Fix broken series blocks when series gets deleted
Fixes #1476

The ability to delete series from Tobira introduced a bug where blocks
referencing that series would be broken. That previous change introduced
the table `all_series` with a `deleted` flag. That way, the foreign key
we had (and is referenced in the deleted comment) became useless. Fixing
this is easy fortunately.