Rename event/series-"status" to -"state"
State was already being used in the database and refers to
the same thing: Being either "ready" or "waiting".
I intruduced the word "status" for this, thinking it was the
technically correct term. But apart from that not even being
necessarily the case (depending on who you ask on the internet),
it's also pretty confusing to mix and use both terms when talking
about the same thing.

That being said, we still use "status" for other things like
workflow activity, and I am not suggesting to change that.