Fix wrong thumbnail icon on series thumbnails
Waiting events were using the audio-only icon in
the "my series" table and the regular video icon
on the "series details" page. They should be using
the "moving-truck" icon, which is quite popular
apparently, but not often seen. So this really
kills two birds with one stone: Fixing the wrong
icon and raising the truck icon's visibility.

Unfortunately this requires a change of a db type and
function, meaning it also needs to recreate two search
views.