Show empty matches on timeline for events with only preview rights
Before, no timeline was shown at all, which was confusing as it's not
clear why the event is even shown in the results. With this commit,
everything is still shown, except for the actual text. Of course this
reveals some information about the transcript, but I don't see a way how
this could get abused or how anyone would consider "knowing a specific
term was used at a specific time" a leak of private data.